I have moved DTOs from Aplicatio-Layer to Shared-Layer and update all namespaces
