+++ Application-Layer/Application-Layer.csproj
@@ -29,6 +29,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Infrastructure-Layer\Infrastructure-Layer.csproj" />
+    <ProjectReference Include="..\Shared-Layer\Shared-Layer.csproj" />
   </ItemGroup>
 
   <ItemGroup>
