+++ Application-Layer/Commands/CourseCommands/UpdateCourse/UpdateCourseCommandValidator.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace Application_Layer.Commands.CourseCommands.UpdateCourse
 {
