+++ Application-Layer/Commands/ModuleCommands/UpdateModule/UpdateModuleCommand.cs
@@ -1,6 +1,6 @@
-﻿using Application_Layer.DTO_s.Module;
-using MediatR;
+﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Shared_Layer.DTO_s.Module;
 
 namespace Application_Layer.Commands.ModuleCommands.UpdateModule
 {
