+++ Application-Layer/Commands/ModuleCommands/UpdateModule/UpdateModuleValidator.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Application_Layer.DTO_s.Module;
-using Application_Layer.Validators.ValidationExtensions;
+﻿using Application_Layer.Validators.ValidationExtensions;
 using FluentValidation;
 
 namespace Application_Layer.Commands.ModuleCommands.UpdateModule
