+++ Application-Layer/Commands/UserCommands/UpdateUser/UpdateUserCommand.cs
@@ -1,6 +1,6 @@
-﻿using Application_Layer.DTO_s;
-using Domain_Layer.Models.User;
+﻿using Domain_Layer.Models.User;
 using MediatR;
+using Shared_Layer.DTO_s.User;
 
 namespace Application_Layer.Commands.UpdateUser
 {
