+++ Test-Layer/CourseTest/IntegrationTest/CourseControllerIntegrationGetCoursesBySearchCriteriaTest.cs
@@ -2,7 +2,6 @@
 using Application_Layer.Queries.CourseQueries.GetAllCoursesBySearchCriteria;
 using Domain_Layer.Models.Course;
 using FakeItEasy;
-using Infrastructure_Layer.Repositories.Course;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
