+++ Test-Layer/CourseTest/IntegrationTest/CourseControllerIntegrationUpdateCourseTest.cs
@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using API_Layer.Controllers;
+﻿using API_Layer.Controllers;
 using Application_Layer.Commands.CourseCommands.UpdateCourse;
-using Application_Layer.DTO_s;
 using Domain_Layer.Models.Course;
 using FakeItEasy;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Shared_Layer.DTO_s.Course;
 
 namespace Test_Layer.CourseTest.IntegrationTest
 {
