+++ Test-Layer/ModuleTest/IntegrationTest/ModuleControllerIntegrationUpdateModuleTest.cs
@@ -1,10 +1,10 @@
 ﻿using API_Layer.Controllers;
 using Application_Layer.Commands.ModuleCommands.UpdateModule;
-using Application_Layer.DTO_s.Module;
 using Domain_Layer.Models.Module;
 using FakeItEasy;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Shared_Layer.DTO_s.Module;
 
 namespace Test_Layer.ModuleTest.IntegrationTest
 {
