+++ Test-Layer/UserTest/IntegrationTests/UserControllerIntegrationLoginTests.cs
@@ -1,11 +1,11 @@
 ﻿using Application_Layer.Commands.RegisterNewUser;
 using Application_Layer.Controllers;
-using Application_Layer.DTO_s;
 using Application_Layer.Queries.LoginUser;
 using Domain_Layer.Models.User;
 using FakeItEasy;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Shared_Layer.DTO_s.User;
 
 namespace Test_Layer.UserTest.IntegrationTests
 {
